fix: round discount_value to 2dp; SCD2 version long (match Polars); assert promo/employee SK 1..N

+++ src/techmart/spark/scd2.py
@@ -13,7 +13,7 @@ def scd2_columns() -> list[SparkColumn]:
         SparkColumn("effective_start_ts", "timestamp", "SCD2 effective start timestamp", nullable=False),
         SparkColumn("effective_end_ts", "timestamp", "SCD2 effective end timestamp; null when current"),
         SparkColumn("is_current", "boolean", "True for the current version of the row", nullable=False),
-        SparkColumn("version", "int", "SCD2 version number (1-based)", nullable=False),
+        SparkColumn("version", "long", "SCD2 version number (1-based)", nullable=False),
     ]
 
 
@@ -24,5 +24,5 @@ def with_scd2_current(df: DataFrame, start: date) -> DataFrame:
         df.withColumn("effective_start_ts", F.lit(start_ts).cast("timestamp"))
         .withColumn("effective_end_ts", F.lit(None).cast("timestamp"))
         .withColumn("is_current", F.lit(True))
-        .withColumn("version", F.lit(1).cast("int"))
+        .withColumn("version", F.lit(1).cast("long"))
     )
